create SecurityController entity User and register form

+++ src/Command/FixturesCommand.php
@@ -5,6 +5,7 @@ namespace App\Command;
 use App\Entity\Question;
 use App\Entity\Subject;
 use App\Entity\Message;
+use App\Entity\User;
 use Doctrine\ORM\EntityManagerInterface;
 use Symfony\Component\Console\Command\Command;
 use Symfony\Component\Console\Input\InputInterface;
@@ -29,20 +30,31 @@ class FixturesCommand extends Command
 
     protected function execute(InputInterface $input, OutputInterface $output)
     {
+        ini_set('memory_limit', -1);
+        ini_set('max_execution_time', 200);
+
         $io = new SymfonyStyle($input, $output);
         $io->text("coucou");
         $io->text("Now loading success");
 
         $faker = \Faker\Factory::create('fr_FR');
 
+        $answer =$io->ask("Truncating all tables... Sure ? yes|no" ,'no');
+        if ($answer !== "yes"){
+            $io->text("Aborting");
+            die();
+        }
+
+
         $conn=$this->em->getConnection();
         //désactive la vérification des clefs etrangères
         $conn->query ('SET FOREIGN_KEY_CHECKS = 0');
-        //écrase les tables
+        //vide les tables
         $conn->query ('TRUNCATE question');
         $conn->query ('TRUNCATE message');
         $conn->query ('TRUNCATE subject');
         $conn->query ('TRUNCATE question_subject');
+        $conn->query ('TRUNCATE user');
         //récupéré la vérification
         $conn->query ('SET FOREIGN_KEY_CHECKS = 1');
 
@@ -80,7 +92,12 @@ class FixturesCommand extends Command
         }
         $this->em->flush();
 
-        for($i=0; $i<20; $i++){
+        // démarre la barre de progression avec 200 opérations
+        $io->progressStart(200);
+
+        for($i=0; $i<100; $i++){
+            //fait avancer la barre de progression
+            $io->progressAdvance(1);
 
             $question = new Question();
 
@@ -97,19 +114,39 @@ class FixturesCommand extends Command
             $question->setStatus($faker->randomElement(['debating', 'voting', 'closed']));
             $question->setSupports($faker->numberBetween(0,4700000));
             $question->setCreationDate($faker->dateTimeBetween ($startDate = '-1 years', $endDate = 'now', $timezone = null) );
-                for($c=0; $c<50;$c++){
-                    $message =new Message();
 
-                    $message->setContent($faker->realText(200));
-
-                    $message->setQuestion($question);
-
-                    //à continuer
-                    $this->em->persist($message);
-
-                }
+            //ajoute des messages sur les questions
+            $messageNumber = mt_rand(0,20);
+            for ($m = 0; $m < $messageNumber; $m++){
+                $message = new Message();
+                $message->setQuestion($question);
+                $message->setClaps($faker->optional(0.5, 0)->numberBetween(0,5000));
+                $message->setCreationDate($faker->dateTimeBetween($question->getCreationDate()));
+                $message->setIsPublished($faker->boolean(95));
+                $message->setContent($faker->paragraphs($nb = mt_rand(1,3), $asText = true));
+                $this->em->persist($message);
+            }
             $this->em->persist($question);
+
+
+
+            // données bidon pour le user
+
+
         }
+
+        for($d=0; $d<100;$d++){
+            $user = new User ();
+            $user->setUsername($faker->name);
+            $user->setEmail($faker->email);
+            $user->setPassword($faker->password);
+            $user->setSocialSecurityNumber($faker->numberBetween(10000,2000000));
+            $user->setRoles($faker->randomElement([['admin'], ['user']]));
+            $this->em->persist($user);
+
+        }
+        // fin de la barre de progression
+        $io->progressFinish();
 
         $this->em->flush();
         $io->success("Done!");
